Erro anterior fixed, basta usar unsigned long long. :+1:

+++ feup-aeda/Date.h
@@ -6,11 +6,11 @@ using namespace std;
 class Date {
 private:
 	unsigned int hour, minutes, day, month, year;
-	unsigned long int compactDate;
+	unsigned long long compactDate;
 public:
 	Date();
 	Date(unsigned int hour, unsigned int minutes, unsigned int day, unsigned int month, unsigned int year);
-	Date(unsigned long int compactDate);
+	Date(unsigned long long compactDate);
 	unsigned int getHour();
 	unsigned int getMinutes();
 	unsigned int getDay();
